Bug 20600: (follow-up) Fix filtering
There was a bug whereby filtering by date would fail if the selected
date syspref was set to dmydot or metric. Rather than use the display
date value, We now use the non-display date value that is always in
a predictable ISO8601 format to construct the Date objects used
for comparison.

Signed-off-by: Owen Leonard <oleonard@myacpl.org>

Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>

Signed-off-by: Nick Clemens <nick@bywatersolutions.com>