Bug 20044: (follow-up) Add min-width to .input_marceditor
This patch adds a min-width property to the first .input_marceditor rule
in order to prevent <selects> styled by Select2 from collapsing at
narrower browser widths.

To reproduce the bug, apply all patches on Bug 20044 except this one.
Open the MARC editor and narrow the browser width to below 768 pixels.
Reload the page and view a tab which contains a <select> styled by
Select2 (942c is often an example). The <select> will have narrowed to
be almost invisible.

Apply the patch and reproduce the steps above. The <select>s should be
fully visible.

Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>
Signed-off-by: Chris Cormack <chrisc@catalyst.net.nz>

Signed-off-by: Nick Clemens <nick@bywatersolutions.com>