Bug 21163: Basket group detail page layout is broken
This fixes a problem with layout of the page when viewing a closed
basket group. Changed the grid options for basket information when a
basket group is closed.

To test, apply the patch and go to Acquisitions -> Vendor -> Basket
groups. You should have multiple baskets and basket groups to test with.

- Edit an open basket group. The layout should look correct, with basket
  group information in the form in the left column and ungrouped basket
  in the right column. Grouping and ungrouping baskets should work
  correctly.
- View a closed basket group. The basket group information should
  display in a single column that spans the main part of the page.

Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>
Signed-off-by: Julian Maurice <julian.maurice@biblibre.com>

Signed-off-by: Nick Clemens <nick@bywatersolutions.com>