Bug 21317: Format long lists under Koha Team tab as columns
This patch adds a couple of new CSS classes for formatting content in
columns: .columns-3 and .columns-4. The four-column class is applied to
the dev team and contributing companies lists on the about page.

The patch also modifies the style of H3 on this page to improve
readability.

Unrelated: "Nelsonville Public Library" corrected to "Athens County
Public Libraries."

To test, apply the patch and regenerate staff client CSS
(https://wiki.koha-community.org/wiki/Working_with_SCSS_in_the_OPAC_and_staff_client).

View the various tabs on the about page and confirm that everything
looks good. Confirm that the dev team and contributing companies lists
are formatted as columns.

Signed-off-by: Martin Renvoize <martin.renvoize@ptfs-europe.com>

Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>

Signed-off-by: Nick Clemens <nick@bywatersolutions.com>