Bug 21380: Circulation history UI improvements - make barcode clickable
This patch modifies the patron circulation history template so that the
barcode of each item links to the item details page.

To test, apply the patch and view the circulation history for a patron
with multiple current or past checkouts.

In the table of titles, the barcode for each item should correctly link
to the item details page for that item.

Signed-off-by: Michal Denar <black23@gmail.com>

Signed-off-by: Josef Moravec <josef.moravec@gmail.com>

Signed-off-by: Nick Clemens <nick@bywatersolutions.com>