Bug 21418: Incorrectly filtered markup in staff client lists
This patch removes the "|html" filter from some variable declarations in
the template used to display the list of lists in the staff client.

To test you should have at least one list. Apply the patch and go to
Lists.

In the table of lists, the "Edit" and "Delete" buttons should look
correct and work correctly.

Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>

Signed-off-by: Josef Moravec <josef.moravec@gmail.com>

Signed-off-by: Nick Clemens <nick@bywatersolutions.com>