Bug 16420: Buttons inconsistent between "Hold found" and "Hold found (waiting)" dialogs in checkin
This patch makes minor changes to the text of hold confirmation buttons
in the checkin process in order to make them more consistent.

Also changed: Removed the "danger" class from the "Ignore" button for
readability and consistency's sake.

To test, apply the patch and check in an item which is on hold for a
patron at your logged-in library. The hold confirmation dialog's buttons
should read "Confirm hold," "Print slip and confirm," and "Ignore."

Confirm the hold, then check in the same item again.

In the dialog, the buttons should read "Confirm hold," "Print slip and
confirm," and "Cancel hold."

Signed-off-by: Michal Denar <black23@gmail.com>
Signed-off-by: Chris Cormack <chrisc@catalyst.net.nz>

Signed-off-by: Nick Clemens <nick@bywatersolutions.com>