Bug 21516: (ALTERNATE) Request notes CSS bug makes them unreadable
This alternate patch makes multiple markup corrections to the ILL
requests templates. It reformats the request details to use a list,
making it consistent with similar interfaces (see patron details,
baskets in acquisitions). The display of notes has been changed so that
it uses a paragraph tag instead of <pre>.

To test, apply the patch and rebuild the staff client CSS.

- Manage an ILL request which has staff and OPAC notes. Everything
  should look readable.
- Running validation on the HTML should only return warnings coming from
  global include files (js_includes.inc).

Signed-off-by: Andrew Isherwood <andrew.isherwood@ptfs-europe.com>

Signed-off-by: Marcel de Rooy <m.de.rooy@rijksmuseum.nl>

Signed-off-by: Nick Clemens <nick@bywatersolutions.com>